kmod/core: Remove stop_machine from kpatch
Now we are able to remove stop_machine safely since we
can count reference threads of target function on-line.

The issue is how to count the threads which is sleep
or running on the target function. The former threads,
we do stack dump on each sleeping(!running) threads,
and the latter one, we do stack dump in the event handler
of __switch_to :)

Note that this still have a issue on unloading.

TODO:
 - Populate monitoring task pool based on # of running tasks.
 - Retry to populate bigger pool if it shorts.
 - Stabilize unpatching feature.

Signed-off-by: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>